+++ CHANGELOG.md
@@ -11,6 +11,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add support of Markdown [#679](https://github.com/nbbrd/spreadsheet4j/issues/679)
 
+### Fixed
+
+- Fix parsing of number format ID in spreadsheet-xl [#693](https://github.com/nbbrd/spreadsheet4j/issues/693)
+
 ## [2.5.10] - 2025-11-17
 
 ### Changed