+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [2.6.0] - 2026-04-23
+
 ### Added
 
 - Add support of Markdown [#679](https://github.com/nbbrd/spreadsheet4j/issues/679)
@@ -210,7 +212,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - JDK11 cleanup
 
-[Unreleased]: https://github.com/nbbrd/spreadsheet4j/compare/v2.5.10...HEAD
+[Unreleased]: https://github.com/nbbrd/spreadsheet4j/compare/v2.6.0...HEAD
+[2.6.0]: https://github.com/nbbrd/spreadsheet4j/compare/v2.5.10...v2.6.0
 [2.5.10]: https://github.com/nbbrd/spreadsheet4j/compare/v2.5.9...v2.5.10
 [2.5.9]: https://github.com/nbbrd/spreadsheet4j/compare/v2.5.8...v2.5.9
 [2.5.8]: https://github.com/nbbrd/spreadsheet4j/compare/v2.5.7...v2.5.8